Extract SharingReportPlugin logic into configureSharingReport [J:MEP-0]
Match the pattern of other plugins: plugin class applies plugins and
delegates to a configure... function in configuration/.
